+++ app/src/main/java/com/boosters/promise/ui/detail/PromiseDetailActivity.kt
@@ -1,15 +1,22 @@
 package com.boosters.promise.ui.detail
 
+import android.content.Intent
 import android.os.Build
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import android.view.Menu
+import android.view.MenuItem
 import androidx.activity.viewModels
+import androidx.appcompat.app.AlertDialog
 import androidx.databinding.DataBindingUtil
 import com.boosters.promise.R
 import com.boosters.promise.data.model.Location
 import com.boosters.promise.databinding.ActivityPromiseDetailBinding
 import com.boosters.promise.ui.detail.adapter.PromiseMemberAdapter
+import com.boosters.promise.ui.promisecalendar.PromiseCalendarActivity
+import com.boosters.promise.ui.promisesetting.PromiseSettingActivity
 import com.boosters.promise.ui.promisesetting.model.PromiseUiState
+import com.google.android.material.snackbar.Snackbar
 import com.naver.maps.geometry.LatLng
 import com.naver.maps.geometry.Tm128
 import com.naver.maps.map.CameraUpdate
@@ -17,7 +24,9 @@ import com.naver.maps.map.MapFragment
 import com.naver.maps.map.NaverMap
 import com.naver.maps.map.OnMapReadyCallback
 import com.naver.maps.map.overlay.Marker
+import dagger.hilt.android.AndroidEntryPoint
 
+@AndroidEntryPoint
 class PromiseDetailActivity : AppCompatActivity(), OnMapReadyCallback {
 
     private lateinit var binding: ActivityPromiseDetailBinding
@@ -30,10 +39,50 @@ class PromiseDetailActivity : AppCompatActivity(), OnMapReadyCallback {
         setBinding()
         initMap()
         setPromiseInfo()
+
+        setSupportActionBar(binding.toolbarPromiseDetail)
+        supportActionBar?.apply {
+            setDisplayShowCustomEnabled(true)
+            setDisplayShowTitleEnabled(false)
+            setDisplayHomeAsUpEnabled(true)
+        }
+
+        promiseDetailViewModel.isDeleted.observe(this) {
+            if (it) {
+                startActivity(
+                    Intent(this, PromiseCalendarActivity::class.java)
+                ).also { finish() }
+                finish()
+            } else {
+                showStateSnackbar(R.string.promiseDetail_delete_ask)
+            }
+        }
     }
 
     override fun onMapReady(map: NaverMap) {
         setObserver(map)
+    }
+
+    override fun onCreateOptionsMenu(menu: Menu?): Boolean {
+        menuInflater.inflate(R.menu.menu_detail_toolbar, menu)
+        return super.onCreateOptionsMenu(menu)
+    }
+
+    override fun onOptionsItemSelected(item: MenuItem): Boolean {
+        when (item.itemId) {
+            R.id.item_detail_edit -> {
+                val intent = Intent(this, PromiseSettingActivity::class.java).putExtra(
+                    PROMISE_INFO_KEY,
+                    promiseDetailViewModel.promiseInfo.value
+                )
+                startActivity(intent)
+                finish()
+            }
+            R.id.item_detail_delete -> {
+                deletePromise()
+            }
+        }
+        return super.onOptionsItemSelected(item)
     }
 
     private fun setBinding() {
@@ -90,6 +139,26 @@ class PromiseDetailActivity : AppCompatActivity(), OnMapReadyCallback {
     private fun convertLocation(location: Location): LatLng {
         val tm128Location = Tm128(location.x.toDouble(), location.y.toDouble())
         return tm128Location.toLatLng()
+    }
+
+    private fun showStateSnackbar(message: Int) {
+        Snackbar.make(binding.root, message, Snackbar.LENGTH_SHORT).show()
+    }
+
+    private fun deletePromise() {
+        AlertDialog.Builder(this)
+            .setMessage(R.string.promiseDetail_delete_ask)
+            .setPositiveButton(R.string.promiseDetail_dialog_yes) { _, _ ->
+                promiseDetailViewModel.removePromise()
+                startActivity(
+                    Intent(this, PromiseCalendarActivity::class.java)
+                ).also { finish() }
+            }
+            .setNegativeButton(R.string.promiseDetail_dialog_no) { _, _ ->
+                return@setNegativeButton
+            }
+            .create()
+            .show()
     }
 
     companion object {
