+++ app/src/main/java/com/boosters/promise/ui/promisesetting/PromiseSettingActivity.kt
@@ -5,7 +5,6 @@ import android.os.Build
 import android.os.Build.VERSION
 import android.os.Bundle
 import android.text.format.DateFormat
-import android.util.Log
 import android.view.MotionEvent
 import android.view.inputmethod.InputMethodManager
 import androidx.activity.result.contract.ActivityResultContracts
