+++ app/src/main/res/layout/activity_promise_detail.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<layout>
+<layout xmlns:tools="http://schemas.android.com/tools">
 
     <data>
 
@@ -10,27 +10,18 @@
 
     <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
         xmlns:app="http://schemas.android.com/apk/res-auto"
-        xmlns:tools="http://schemas.android.com/tools"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         tools:context=".ui.detail.PromiseDetailActivity">
 
-        <com.google.android.material.appbar.AppBarLayout
-            android:id="@+id/appBarLayout_promiseDetail_appBar"
+        <androidx.appcompat.widget.Toolbar
+            android:id="@+id/toolbar_promiseDetail"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             app:elevation="0dp"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent">
-
-            <com.google.android.material.appbar.MaterialToolbar
-                android:id="@+id/appBar_promiseDetail_appBar"
-                android:layout_width="match_parent"
-                android:layout_height="?attr/actionBarSize"
-                android:theme="@style/Promise.ActionBar.TopAppBar"
-                app:menu="@menu/menu_detail_appbar" />
-        </com.google.android.material.appbar.AppBarLayout>
+            app:layout_constraintTop_toTopOf="parent" />
 
         <androidx.constraintlayout.widget.Guideline
             android:id="@+id/guideline_promiseDetail_left"
@@ -63,7 +54,7 @@
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintHorizontal_bias="0"
             app:layout_constraintStart_toStartOf="@+id/guideline_promiseDetail_left"
-            app:layout_constraintTop_toBottomOf="@+id/appBarLayout_promiseDetail_appBar"
+            app:layout_constraintTop_toBottomOf="@+id/toolbar_promiseDetail"
             app:layout_constraintVertical_bias="0"
             app:layout_constraintVertical_chainStyle="packed"
             tools:text="부스터즈 모각코" />
@@ -134,7 +125,8 @@
             app:layout_constraintEnd_toStartOf="@+id/guideline_promiseDetail_right"
             app:layout_constraintStart_toStartOf="@+id/guideline_promiseDetail_left"
             app:layout_constraintTop_toBottomOf="@+id/textView_promiseDetail_date"
-            tools:background="@color/azure_radiance" />
+            tools:background="@color/azure_radiance"
+            tools:layout="@layout/activity_promise_detail" />
 
         <androidx.recyclerview.widget.RecyclerView
             android:id="@+id/recyclerView_promiseDetail_memberList"
